Warden security audit and hardening
